a little org change
also an experiment on whether this kind of modification works

this is before starting to do anything with the starter project,
so perhaps is the easiest stage to commit such a change.

i'd imagine this kind of change will get more and more difficult
as the code is developed and becomes more complex.