presentation
- use loggedInUser (moved to 'global
scope' to be accessible by other
widgets) to check for user's email
and conditionally style message
bubbles.

- also manage order of ListView
items rendered. I am still not sure
using that reversed on two things
is the best way to achieve this.

TODO: have to figure out sorting
messages chronologically, and it
seems that's not happening unless
we add timestamps to the data
(couldn't find any other way)